build!: migrate from maven to gradle
This makes it much easier to build in CI, as there's no need to build
Paper seperately and copy the re-obfuscated server jar file into the
lib/ folder for Maven to load. Additionally, there's no reason for me to
guess what each obfuscated class does, it's all remapped from the
deobfuscated mappings to the obfuscated ones in the reobfJar task!