fix: use assets if they end in .jar & are streams
Apparently, something went wrong with someone/something's build system
somewhere and their Jar file came out as application/octet-stream and
not application/java-archive. This could be due to a multitude of
things, however since it only affects us and not the average Joe
dropping Jar files into his plugins/ folder, we have to patch it.